requirements.txt
pandas
python-dotenv
colorama